# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Clinic API",
    "version": "2.0.0",
    "description": "واجهة REST لنظام إدارة عيادة التغذية. مصمَّمة ليستهلكها موقع الويب وتطبيق الموبايل (Android/iOS) الخاص بالمريض دون أي تعديل. المصادقة: POST /api/auth/login ثم إرسال `Authorization: Bearer <token>` في كل طلب. التوكن صالح 12 ساعة.",
    "contact": {
      "name": "Clinic IT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "تطوير محلي"
    },
    {
      "url": "https://api.clinic.example",
      "description": "الإنتاج"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "قائمة + بحث سريع",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "اسم أو رقم ملف أو هاتف"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "archived"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "file_no",
                "name",
                "created"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ total, items: Patient[] }"
          }
        }
      },
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "إنشاء مريض (رقم الملف تلقائي)",
        "responses": {
          "201": {
            "description": "Patient"
          },
          "400": {
            "description": "تحقق المدخلات"
          }
        }
      }
    },
    "/api/patients/{id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "مريض واحد",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Patient"
          },
          "404": {
            "description": "غير موجود"
          }
        }
      },
      "put": {
        "tags": [
          "patients"
        ],
        "summary": "تحديث جزئي أو كامل",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Patient"
          }
        }
      },
      "delete": {
        "tags": [
          "patients"
        ],
        "summary": "حذف (admin) أو أرشفة لغير المدير",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "{ deleted | archived }"
          },
          "403": {
            "description": "صلاحية"
          }
        }
      }
    },
...
    "/api/reports/patient/{id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "تقرير كامل للمريض (للطباعة/PDF)",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "PatientReport"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    },
...
  }
}